% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridworld.R
\name{gridworldEnvironment}
\alias{gridworldEnvironment}
\title{Defines an environment for a gridworld example}
\usage{
gridworldEnvironment(state, action)
}
\arguments{
\item{state}{The current state.}

\item{action}{Action to be executed.}
}
\value{
List containing the next state and the reward.
}
\description{
Function defines an environment for a 2x2 gridworld example. Here, an agent is intended to
navigate from an arbitrary start position to a goal position. The grid is surrounded by a wall
which makes it impossible for the agent to move off the grid. In addition, the agent faces a wall between s1 and s4.
If the agent reaches the goal position, it earns a reward of 10. Crossing each square of the grid leads
to a negative reward of -1.
}

