% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionSelection.R
\name{lookupActionSelection}
\alias{lookupActionSelection}
\title{Converts a name into an action selection function}
\usage{
lookupActionSelection(type)
}
\arguments{
\item{type}{A string denoting the type of action selection. Allowed values are \code{epsilon-greedy} or \code{random}.}
}
\value{
Function that implements the specific learning rule.
}
\description{
Input is a name for the action selection, output is the corresponding function object.
}

