% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningRule.R
\name{lookupLearningRule}
\alias{lookupLearningRule}
\title{Loads reinforcement learning algorithm}
\usage{
lookupLearningRule(type)
}
\arguments{
\item{type}{A string denoting the learning rule. Allowed values are \code{experienceReplay}.}
}
\value{
Function that implements the specific learning rule.
}
\description{
Decides upon a learning rule for reinforcement learning.
Input is a name for the learning rule, while output is the corresponding function object.
}
\seealso{
\code{\link{ReinforcementLearning}}
}
