% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{policy}
\alias{policy}
\title{Calculates the reinforcement learning policy}
\usage{
policy(x)
}
\arguments{
\item{x}{Variable which encodes the behavior of the agent. This can be
either a \code{matrix}, \code{data.frame} or an \code{\link{rl}} object.}
}
\value{
Returns the learned policy.
}
\description{
Calculates reinforcement learning policy from a given state-action table Q.
The policy is the decision-making function of the agent and defines the learning
agent's behavior at a given time.
}
\seealso{
\code{\link{ReinforcementLearning}}
}

