% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionSelection.R
\name{randomActionSelection}
\alias{randomActionSelection}
\title{Performs random action selection}
\usage{
randomActionSelection(Q, state, epsilon)
}
\arguments{
\item{Q}{State-action table of type \code{hash}.}

\item{state}{The current state.}

\item{epsilon}{Exploration rate between 0 and 1 (not used).}
}
\value{
Character value defining the next action.
}
\description{
Performs random action selection. In this strategy, the agent always selects an action at random.
}

