% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state}
\alias{state}
\title{Creates a state representation for arbitrary objects}
\usage{
state(x, ...)
}
\arguments{
\item{x}{An object of any class.}

\item{...}{Additional parameters passed to function.}
}
\value{
Character value defining the state representation of the given object.
}
\description{
Converts object of any class to a reinforcement learning state of type character.
}

