% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{policy}
\alias{policy}
\title{Computes the reinforcement learning policy}
\usage{
policy(x)
}
\arguments{
\item{x}{Variable which encodes the behavior of the agent. This can be
either a \code{matrix}, \code{data.frame} or an \code{\link{rl}} object.}
}
\value{
Returns the learned policy.
}
\description{
Deprecated. Please use [ReinforcementLearning::computePolicy()] instead.
}
\seealso{
\code{\link{ReinforcementLearning}}
}
